refactor: 백엔드 패키지 구조 정리
- 루트 고아 패키지(controller/service/dto) 해소: announcement, healthcheck 패키지로 이동
- TextNormalizer를 infra/ai/service에서 common/util로 이동
- ProfileImageService, MemberRedisSyncer, ProfileImageProperties를 auth에서 member 패키지로 이동
- GameConstants는 도메인 계층 유지 (domain → application 역방향 의존 방지)
