+++ backend/src/main/java/com/gakkaweo/backend/admin/controller/AdminUserController.java
@@ -8,8 +8,8 @@ import com.gakkaweo.backend.admin.dto.UserGameHistoryResponse;
 import com.gakkaweo.backend.admin.dto.UserListResponse;
 import com.gakkaweo.backend.admin.service.AdminUserService;
 import com.gakkaweo.backend.auth.security.CustomUserDetails;
-import com.gakkaweo.backend.auth.service.MemberRedisSyncer;
 import com.gakkaweo.backend.config.openapi.AdminErrorResponses;
+import com.gakkaweo.backend.member.service.MemberRedisSyncer;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
