+++ backend/src/main/java/com/gakkaweo/backend/admin/service/AdminUserService.java
@@ -11,8 +11,6 @@ import com.gakkaweo.backend.admin.dto.UserListResponse;
 import com.gakkaweo.backend.admin.sort.SortRequestParser;
 import com.gakkaweo.backend.admin.sort.SortSpecBuilder;
 import com.gakkaweo.backend.admin.sort.UserSortField;
-import com.gakkaweo.backend.auth.service.MemberRedisSyncer;
-import com.gakkaweo.backend.auth.service.ProfileImageService;
 import com.gakkaweo.backend.common.exception.BusinessException;
 import com.gakkaweo.backend.common.exception.ErrorCode;
 import com.gakkaweo.backend.common.util.EnumParser;
@@ -27,6 +25,8 @@ import com.gakkaweo.backend.domain.member.repository.LocalAccountRepository;
 import com.gakkaweo.backend.domain.member.repository.MemberRepository;
 import com.gakkaweo.backend.domain.member.repository.SocialAccountRepository;
 import com.gakkaweo.backend.domain.member.validation.NicknameValidator;
+import com.gakkaweo.backend.member.service.MemberRedisSyncer;
+import com.gakkaweo.backend.member.service.ProfileImageService;
 import com.gakkaweo.backend.ranking.event.RankingUpdateEvent;
 import com.gakkaweo.backend.ranking.service.RankingService;
 import jakarta.persistence.criteria.Predicate;
