+++ backend/src/main/java/com/gakkaweo/backend/auth/controller/AuthController.java
@@ -9,11 +9,11 @@ import com.gakkaweo.backend.auth.security.CustomUserDetails;
 import com.gakkaweo.backend.auth.service.AccountService;
 import com.gakkaweo.backend.auth.service.AuthService;
 import com.gakkaweo.backend.auth.service.LocalAuthService;
-import com.gakkaweo.backend.auth.service.MemberRedisSyncer;
-import com.gakkaweo.backend.auth.service.ProfileImageService;
 import com.gakkaweo.backend.auth.util.CookieUtils;
 import com.gakkaweo.backend.config.openapi.StandardErrorResponses;
 import com.gakkaweo.backend.domain.member.entity.Member;
+import com.gakkaweo.backend.member.service.MemberRedisSyncer;
+import com.gakkaweo.backend.member.service.ProfileImageService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.headers.Header;
