+++ backend/src/test/java/com/gakkaweo/backend/auth/AccountDeletionTest.java
@@ -2,13 +2,13 @@ package com.gakkaweo.backend.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.gakkaweo.backend.auth.config.ProfileImageProperties;
 import com.gakkaweo.backend.auth.dto.AuthResponse;
 import com.gakkaweo.backend.auth.dto.TokenPair;
 import com.gakkaweo.backend.domain.auth.repository.RefreshTokenRepository;
 import com.gakkaweo.backend.domain.member.entity.Member;
 import com.gakkaweo.backend.domain.member.repository.LocalAccountRepository;
 import com.gakkaweo.backend.domain.member.repository.MemberRepository;
+import com.gakkaweo.backend.member.config.ProfileImageProperties;
 import com.gakkaweo.backend.support.IntegrationTestBase;
 import java.io.IOException;
 import java.nio.file.Files;
